Chain interceptors (#79)
This provides methods for intercepting client conns and servers that
allow the user to provide a (var-arg) chain of interceptors. So if a
user wants to apply multiple interceptors, this greatly simplifies
usage.

The tests are kinda complicated, but they are basically asserting that
every interceptor gets run once and in the expected order. And they test
a variety of "setup" cases -- one that uses the chain to add an entire
batch of interceptors at once and one that adds them one-at-a-time, etc
-- to make sure they behave the same in all cases.

Resolves #72.